Added find method for better searching of record types coming back from the array processor.
What I thought was once 'normalization' i now realize is just arbitrary normalization, and in reality the tab indentation depends entirely on the length of the alias that leads the record. This will dictate even further investigation into how to better process records coming from the named-checkzone utility - there has to be a logic template for parsing these values, and then reinterpreting them. I can't always assume that a value will be in a position. If someone says 'supercalifragalisticexpiealadocious' is their subdomain - this will turn into chunky salsa. Warrants pinning and coming back to re-evaluate